software/requirements.txt
Pillow==10.3.0
numpy==1.26.4
Flask==3.0.3
uuid==1.30
Flask-Cors==4.0.0